ReFormato para Terminal

+++ src/main/java/com/nocolau/termoservidor/Controlador.java
@@ -35,12 +35,12 @@ public class Controlador {
                     nConfig.setPuerto(Integer.parseInt(args[0]));
                     break;
                 case 2:
-                    tiempoXMuestra = (int)(Long.parseLong(args[0]) / Integer.parseInt(args[1]));
+                    tiempoXMuestra = (int)(Long.parseLong(args[0]) / Float.parseFloat(args[1]));
                     nConfig.setTIEMPO_BUCLE(Long.parseLong(args[0])* 60000);
                     nConfig.setTIEMPO_DIV_VARIABLE(tiempoXMuestra);
                     break;
                 case 3:
-                    tiempoXMuestra = (int)(Long.parseLong(args[1]) / Integer.parseInt(args[2]));
+                    tiempoXMuestra = (int)(Long.parseLong(args[1]) / Float.parseFloat(args[2]));
                     nConfig.setPuerto(Integer.parseInt(args[0]));
                     nConfig.setTIEMPO_BUCLE(Long.parseLong(args[1])* 60000);
                     nConfig.setTIEMPO_DIV_VARIABLE(tiempoXMuestra);
